deleted the partial json parser moved jni setup to the main package
