+++ real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/main/JNISetup.java
@@ -1,4 +1,4 @@
-package com.roboclub.robobuggy.utilities;
+package com.roboclub.robobuggy.main;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
