+++ real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/main/RobobuggyMainFile.java
@@ -3,7 +3,6 @@ package com.roboclub.robobuggy.main;
 import com.roboclub.robobuggy.robots.AbstractRobot;
 import com.roboclub.robobuggy.robots.TransistorDataCollection;
 import com.roboclub.robobuggy.ui.Gui;
-import com.roboclub.robobuggy.utilities.JNISetup;
 
 
 /**
